Checkstyle: Rename g.s.engine.gamePlayer -> g.s.engine.player
