+++ game-core/src/main/java/games/strategy/engine/delegate/DefaultDelegateBridge.java
@@ -11,9 +11,9 @@ import games.strategy.engine.display.IDisplay;
 import games.strategy.engine.framework.AbstractGame;
 import games.strategy.engine.framework.IGame;
 import games.strategy.engine.framework.ServerGame;
-import games.strategy.engine.gamePlayer.IRemotePlayer;
 import games.strategy.engine.history.IDelegateHistoryWriter;
 import games.strategy.engine.message.MessengerException;
+import games.strategy.engine.player.IRemotePlayer;
 import games.strategy.engine.random.IRandomSource;
 import games.strategy.engine.random.IRandomStats.DiceType;
 import games.strategy.engine.random.RandomStats;
