+++ game-core/src/main/java/games/strategy/engine/framework/AbstractGame.java
@@ -12,12 +12,12 @@ import games.strategy.engine.data.PlayerList;
 import games.strategy.engine.data.PlayerManager;
 import games.strategy.engine.data.events.GameStepListener;
 import games.strategy.engine.display.IDisplay;
-import games.strategy.engine.gamePlayer.DefaultPlayerBridge;
-import games.strategy.engine.gamePlayer.IGamePlayer;
-import games.strategy.engine.gamePlayer.IPlayerBridge;
 import games.strategy.engine.message.IChannelMessenger;
 import games.strategy.engine.message.IRemoteMessenger;
 import games.strategy.engine.message.RemoteName;
+import games.strategy.engine.player.DefaultPlayerBridge;
+import games.strategy.engine.player.IGamePlayer;
+import games.strategy.engine.player.IPlayerBridge;
 import games.strategy.engine.vault.Vault;
 import games.strategy.net.IMessenger;
 import games.strategy.net.INode;
