+++ game-core/src/main/java/games/strategy/engine/framework/ClientGame.java
@@ -10,9 +10,9 @@ import java.util.logging.Level;
 import games.strategy.engine.data.Change;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
-import games.strategy.engine.gamePlayer.IGamePlayer;
 import games.strategy.engine.history.EventChild;
 import games.strategy.engine.message.RemoteName;
+import games.strategy.engine.player.IGamePlayer;
 import games.strategy.engine.random.IRandomSource;
 import games.strategy.engine.random.IRemoteRandom;
 import games.strategy.engine.random.RemoteRandom;
