+++ game-core/src/main/java/games/strategy/engine/framework/startup/mc/ClientModel.java
@@ -49,13 +49,13 @@ import games.strategy.engine.framework.startup.ui.ClientOptions;
 import games.strategy.engine.framework.startup.ui.PlayerType;
 import games.strategy.engine.framework.ui.SaveGameFileChooser;
 import games.strategy.engine.framework.ui.background.WaitWindow;
-import games.strategy.engine.gamePlayer.IGamePlayer;
 import games.strategy.engine.message.ChannelMessenger;
 import games.strategy.engine.message.IChannelMessenger;
 import games.strategy.engine.message.IRemoteMessenger;
 import games.strategy.engine.message.RemoteMessenger;
 import games.strategy.engine.message.RemoteName;
 import games.strategy.engine.message.unifiedmessenger.UnifiedMessenger;
+import games.strategy.engine.player.IGamePlayer;
 import games.strategy.io.IoUtils;
 import games.strategy.net.ClientMessenger;
 import games.strategy.net.CouldNotLogInException;
