+++ game-core/src/main/java/games/strategy/engine/player/DefaultPlayerBridge.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.gamePlayer;
+package games.strategy.engine.player;
 
 import static com.google.common.base.Preconditions.checkNotNull;
