+++ game-core/src/main/java/games/strategy/engine/player/IRemotePlayer.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.gamePlayer;
+package games.strategy.engine.player;
 
 import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.message.IRemote;
