+++ game-core/src/main/java/games/strategy/triplea/delegate/AbstractDelegate.java
@@ -7,7 +7,7 @@ import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.delegate.IDelegate;
 import games.strategy.engine.delegate.IDelegateBridge;
 import games.strategy.engine.display.IDisplay;
-import games.strategy.engine.gamePlayer.IRemotePlayer;
+import games.strategy.engine.player.IRemotePlayer;
 import games.strategy.sound.ISound;
 
 /**
