+++ game-core/src/main/java/games/strategy/triplea/ui/BattlePanel.java
@@ -30,7 +30,7 @@ import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.data.Territory;
 import games.strategy.engine.data.Unit;
 import games.strategy.engine.framework.GameRunner;
-import games.strategy.engine.gamePlayer.IGamePlayer;
+import games.strategy.engine.player.IGamePlayer;
 import games.strategy.engine.random.PbemDiceRoller;
 import games.strategy.net.GUID;
 import games.strategy.triplea.TripleAPlayer;
