+++ game-core/src/main/java/games/strategy/triplea/ui/EndTurnPanel.java
@@ -4,8 +4,8 @@ import javax.swing.Action;
 import javax.swing.JOptionPane;
 
 import games.strategy.engine.data.GameData;
-import games.strategy.engine.gamePlayer.IPlayerBridge;
 import games.strategy.engine.pbem.ForumPosterComponent;
+import games.strategy.engine.player.IPlayerBridge;
 import games.strategy.triplea.delegate.remote.IAbstractForumPosterDelegate;
 import games.strategy.ui.SwingAction;
